Owner added to City+Region+MustSee at creation time + must be logged in and owner to delete a City/Region/MustSee

+++ must_see/views.py
@@ -4,9 +4,12 @@ from rest_framework import status
 from .models import MustSee
 from .serializers.common import MustSerializer
 from lib.exceptions import exceptions
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.exceptions import PermissionDenied
 
 
 class MustSeeListView(APIView):
+    permission_classes = (IsAuthenticatedOrReadOnly, )
     # ENDPOINT GET /api/mustsee/
     @exceptions
     def get(self, request):
@@ -15,9 +18,10 @@ class MustSeeListView(APIView):
         return Response(serialized_musts.data)
     
     # ENDPOINT POST /api/mustsee/
+    # -- OWNER IS ADDED TO ENTRY --
     @exceptions
     def post(self, request):
-        must_to_create = MustSerializer(data=request.data)
+        must_to_create = MustSerializer(data={ **request.data, 'owner': request.user.id })
         must_to_create.is_valid(raise_exception=True)
         must_to_create.save()
         return Response(must_to_create.data, status.HTTP_201_CREATED)
@@ -33,9 +37,12 @@ class MustSingleView(APIView):
     
 
     # ENDPOINT PUT /api/mustsee/<pk>/
+    # -- ONLY OWNER CAN UPDATE MUSTSEE --
     @exceptions
     def put(self, request, pk):
         must_to_update = MustSee.objects.get(pk=pk)
+        if must_to_update.owner != request.user and not request.user.is_staff:
+            raise PermissionDenied()
         serialized_must_to_update = MustSerializer(must_to_update, request.data, partial=True)
         serialized_must_to_update.is_valid(raise_exception=True)
         serialized_must_to_update.save()
@@ -43,8 +50,11 @@ class MustSingleView(APIView):
     
 
     # ENDPOINT DELETE /api/mustsee/<pk>/
+    # -- ONLY OWNER CAN DELETE MUSTSEE --
     @exceptions
     def delete(self, request, pk):
         must_to_delete = MustSee.objects.get(pk=pk)
+        if must_to_delete.owner != request.user and not request.user.is_staff:
+            raise PermissionDenied()
         must_to_delete.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
