+++ regions/views.py
@@ -4,9 +4,12 @@ from rest_framework import status
 from .models import Region
 from .serializers.common import RegionSerializer
 from lib.exceptions import exceptions
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.exceptions import PermissionDenied
 
 
 class RegionsListView(APIView):
+    permission_classes = (IsAuthenticatedOrReadOnly, )
     # ENDPOINT GET /api/regions/
     @exceptions
     def get(self, request):
@@ -15,9 +18,10 @@ class RegionsListView(APIView):
         return Response(serialized_regions.data)
     
     # ENDPOINT POST /api/regions/
+    # -- OWNER IS ADDED TO ENTRY --
     @exceptions
     def post(self, request):
-        region_to_create = RegionSerializer(data=request.data)
+        region_to_create = RegionSerializer(data={ **request.data, 'owner': request.user.id })
         region_to_create.is_valid(raise_exception=True)
         region_to_create.save()
         return Response(region_to_create.data, status.HTTP_201_CREATED)
@@ -33,9 +37,12 @@ class RegionSingleView(APIView):
     
 
     # ENDPOINT PUT /api/regions/<pk>/
+    # -- ONLY OWNER CAN UPDATE REGION --
     @exceptions
     def put(self, request, pk):
         region_to_update = Region.objects.get(pk=pk)
+        if region_to_update.owner != request.user and not request.user.is_staff:
+            raise PermissionDenied()
         serialized_region_to_update = RegionSerializer(region_to_update, request.data, partial=True)
         serialized_region_to_update.is_valid(raise_exception=True)
         serialized_region_to_update.save()
@@ -43,8 +50,11 @@ class RegionSingleView(APIView):
     
 
     # ENDPOINT DELETE /api/regions/<pk>/
+    # -- ONLY OWNER CAN DELETE REGION --
     @exceptions
     def delete(self, request, pk):
         region_to_delete = Region.objects.get(pk=pk)
+        if region_to_delete.owner != request.user and not request.user.is_staff:
+            raise PermissionDenied()
         region_to_delete.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
